fix: add links directly to this.state.data
This is counter-intuitive. However, it works because we are modifying an existing array. Problems are caused when more sections are loaded but links haven't loaded for previous sections. in this case, if we used a closure for this.state.data we would inadvertently delete a new section that was added.